dump-documents command working
